Extract ConvertFeatures to a separate pkg
