+++ api/v1/imagescan/convert.go
@@ -4,7 +4,6 @@ import (
 	"strings"
 
 	"github.com/pkg/errors"
-	log "github.com/sirupsen/logrus"
 	"github.com/stackrox/rox/pkg/utils"
 	apiV1 "github.com/stackrox/scanner/api/v1"
 	"github.com/stackrox/scanner/api/v1/common"
@@ -38,51 +37,6 @@ var (
 		return m
 	}()
 )
-
-func convertVulnerabilities(apiVulns []apiV1.Vulnerability) []*v1.Vulnerability {
-	vulns := make([]*v1.Vulnerability, 0, len(apiVulns))
-	for _, v := range apiVulns {
-		metadata, err := convert.MetadataMap(v.Metadata)
-		if err != nil {
-			log.Errorf("error converting metadata map: %v", err)
-			continue
-		}
-		if metadata == nil {
-			log.Warnf("metadata is nil for %s; skipping...", v.Name)
-			continue
-		}
-
-		vulns = append(vulns, &v1.Vulnerability{
-			Name:        v.Name,
-			Description: v.Description,
-			Link:        v.Link,
-			MetadataV2:  metadata,
-			FixedBy:     v.FixedBy,
-			Severity:    v.Severity,
-		})
-	}
-	return vulns
-}
-
-// ConvertFeatures converts api Features into v1 (proto) Feature pointers.
-func ConvertFeatures(apiFeatures []apiV1.Feature) []*v1.Feature {
-	features := make([]*v1.Feature, 0, len(apiFeatures))
-	for _, a := range apiFeatures {
-		vulns := convertVulnerabilities(a.Vulnerabilities)
-
-		features = append(features, &v1.Feature{
-			Name:                a.Name,
-			Version:             a.Version,
-			FeatureType:         a.VersionFormat,
-			AddedByLayer:        a.AddedBy,
-			Location:            a.Location,
-			Vulnerabilities:     vulns,
-			FixedBy:             a.FixedBy,
-			ProvidedExecutables: a.Executables,
-		})
-	}
-	return features
-}
 
 func convertLanguageLevelComponents(layersToComponents []*component.LayerToComponents) map[string]*v1.LanguageLevelComponents {
 	converted := make(map[string]*v1.LanguageLevelComponents, len(layersToComponents))
